split ebt logic into separate modules
