+++ solar/src/actors/replication/ebt/mod.rs
@@ -1,0 +1,6 @@
+mod clock;
+mod manager;
+mod replicator;
+
+pub use clock::{EncodedClockValue, VectorClock};
+pub use manager::{EbtEvent, EbtManager, SessionRole};
